requirements.txt
altgraph==0.17.4
blinker==1.9.0
click==8.2.1
Flask==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
packaging==25.0
pycryptodome==3.23.0
pyinstaller==6.14.1
pyinstaller-hooks-contrib==2025.5
setuptools==80.9.0
Werkzeug==3.1.3
